feat: api loading優化，新增apiCount處理多支api發送情況
